refactor : 성공/실패 모달창 통일로 인한 클래스명 수정
